requirements.txt
# Requirements for CliniQ AI v2.0
# Python version: 3.11.5

# Core Flask packages
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Security==3.0.0
Flask-Session==0.5.0

# Machine Learning and AI packages
tensorflow==2.13.0
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
scikit-learn==1.3.0
scikit-image==0.21.0
joblib==1.3.2

# Image processing
Pillow==10.0.0

# Utility packages
Werkzeug==2.3.7
email-validator==2.0.0
natsort==8.4.0

# Roboflow for ECG detection
roboflow==1.1.7
supervision==0.14.0


gunicorn
